chore: refactor auth and permission based logic
