+++ src/modules/auth/auth.db.ts
@@ -13,3 +13,37 @@ export function getSecretKeybyKeyId(
 
   return pgClient.query(query, [keyId, appPid]);
 }
+
+export function getUserByClientId(pgClient: PoolClient, clientId: string): Promise<QueryResult> {
+  const query = `
+    SELECT 
+      id, 
+      "appId", 
+      "blockedAt", 
+      "clientId", 
+      "createdAt", 
+      "firstName", 
+      "lastName", 
+      "isOnline", 
+      "lastOnline", 
+      "orgId", 
+      "updatedAt", 
+      "username"
+    FROM authentication_users
+    WHERE "clientId" = $1;
+  `;
+
+  return pgClient.query(query, [clientId]);
+}
+
+// export function getPermissionsByKeyId(pgClient: PoolClient, keyId: string): Promise<QueryResult> {
+//   const query = `
+//     SELECT cpe.permission, cpa.pattern
+//     FROM credentials c
+//     LEFT JOIN credential_permissions cpe ON cpe."credentialId" = c.id
+//     LEFT JOIN credential_patterns cpa ON cpa."credentialId" = c.id
+//     WHERE c."keyId" = $1;
+//   `;
+
+//   return pgClient.query(query, [keyId]);
+// }
